backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
gunicorn==23.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
boto3==1.40.46
botocore==1.40.46
livekit==1.0.16
livekit-agents==1.2.14
livekit-api==1.0.6
livekit-protocol==1.0.7
livekit-plugins-deepgram==1.2.14
livekit-plugins-openai==1.2.14
livekit-plugins-sarvam==1.2.14
livekit-plugins-silero==1.2.14
livekit-plugins-turn-detector==1.2.14
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-elevenlabs==1.2.14
livekit-blingfire==1.0.0
livekit-plugins-aws
sounddevice==0.5.2
av==15.1.0
onnxruntime==1.23.0
qdrant-client==1.9.0
faiss-cpu==1.12.0
fastmcp==2.13.0
pymongo==4.6.3
motor==3.3.2
httpx==0.28.1
aiohttp==3.13.0
requests==2.32.5
pandas==2.3.3
PyYAML==6.0.3
aiofiles==24.1.0
pypdf2==3.0.1
python-docx==1.1.0

